+++ tests/test_external_commands.py
@@ -14,15 +14,8 @@ from unittest.mock import patch
 import pytest
 from loguru import logger
 
-from bin.processes import (
-    run_aragorn,
-    run_mash_sketch,
-    run_minced,
-    run_phanotate,
-    run_pyrodigal,
-    run_pyrodigal_gv,
-)
-
+from bin.processes import (run_aragorn, run_mash_sketch, run_minced,
+                           run_phanotate, run_pyrodigal, run_pyrodigal_gv)
 # import functions
 from bin.util import remove_directory