+++ tests/test_proteins.py
@@ -8,7 +8,6 @@ Usage: pytest .
 # import
 import os
 import shutil
-
 # import functions
 import subprocess
 import sys